+++ lib/better_translate.rb
@@ -2,9 +2,11 @@
 require "ruby-progressbar"
 
 require "better_translate/version"
+require "better_translate/utils"
 require "better_translate/translator"
 require "better_translate/service"
 require "better_translate/writer"
+require "better_translate/helper"
 
 require 'better_translate/providers/base_provider'
 require 'better_translate/providers/chatgpt_provider'
@@ -23,7 +25,8 @@ module BetterTranslate
     # Metodo install per generare il file di configurazione (initializer)
     def install
       unless defined?(Rails) && Rails.respond_to?(:root)
-        puts "Il metodo install è disponibile solo in un'applicazione Rails."
+        message = "Il metodo install è disponibile solo in un'applicazione Rails."
+        BetterTranslate::Utils.logger(message: message)
         return
       end
 
@@ -32,19 +35,23 @@ module BetterTranslate
       destination = File.join(Rails.root, "config", "initializers", "better_translate.rb")
 
       if File.exist?(destination)
-        puts "Il file initializer esiste già: #{destination}"
+        message = "Il file initializer esiste già: #{destination}"
+        BetterTranslate::Utils.logger(message: message)
       else
         FileUtils.mkdir_p(File.dirname(destination))
         FileUtils.cp(source, destination)
-        puts "Initializer creato in: #{destination}"
+        message = "Il file initializer esiste già: #{destination}"
+        BetterTranslate::Utils.logger(message: message)
       end
     end
 
     def magic
-      puts "Metodo magic invocato: eseguirò la traduzione dei file..."
+      message = "Metodo magic invocato: eseguirò la traduzione dei file..."
+      BetterTranslate::Utils.logger(message: message)
 
       BetterTranslate::Translator.work
-      puts "Metodo magic invocato: Traduzione completata con successo!"
+      message = "Metodo magic invocato: Traduzione completata con successo!"
+      BetterTranslate::Utils.logger(message: message)
     end
 
   end